requirements.txt
streamlit
requests
python-docx
beautifulsoup4
python-dotenv
groq
difflib
